Remove feature check for update button and consolidate prepublish checking for publish and update buttons
