+++ assets/src/edit-story/components/header/buttons/publish.js
@@ -32,13 +32,9 @@ import { toDate, isAfter, subMinutes, getOptions } from '../../../../date';
 import { TRANSITION_DURATION } from '../../dialog';
 import { useStory, useLocalMedia, useConfig } from '../../../app';
 import useRefreshPostEditURL from '../../../utils/useRefreshPostEditURL';
-import { Primary } from '../../button';
-import WithTooltip from '../../tooltip';
 import TitleMissingDialog from '../titleMissingDialog';
 import useHeader from '../use';
-import { usePrepublishChecklist } from '../../inspector/prepublish';
-import { PRE_PUBLISH_MESSAGE_TYPES } from '../../../app/prepublish';
-import { ButtonContent, WarningIcon } from './styles';
+import ButtonWithChecklistWarning from './buttonWithChecklistWarning';
 
 function Publish() {
   const { isSaving, date, storyId, saveStory, title } = useStory(
@@ -56,14 +52,6 @@ function Publish() {
   const { titleInput } = useHeader();
   const [showDialog, setShowDialog] = useState(false);
   const { capabilities } = useConfig();
-
-  const { checklist, refreshChecklist } = usePrepublishChecklist();
-
-  const tooltip = checklist.some(
-    ({ type }) => PRE_PUBLISH_MESSAGE_TYPES.ERROR === type
-  )
-    ? __('There are items in the checklist to resolve', 'web-stories')
-    : null;
 
   const refreshPostEditURL = useRefreshPostEditURL(storyId);
   // Offset the date by one minute to accommodate for network latency.
@@ -105,26 +93,13 @@ function Publish() {
     ? __('Schedule', 'web-stories')
     : __('Publish', 'web-stories');
 
-  const button = (
-    <Primary
-      onPointerEnter={refreshChecklist}
-      onClick={handlePublish}
-      isDisabled={!capabilities?.hasPublishAction || isSaving || isUploading}
-    >
-      <ButtonContent>
-        {text}
-        {tooltip && <WarningIcon />}
-      </ButtonContent>
-    </Primary>
-  );
-
-  const wrappedWithTooltip = (
-    <WithTooltip title={tooltip}>{button}</WithTooltip>
-  );
-
   return (
     <>
-      {tooltip ? wrappedWithTooltip : button}
+      <ButtonWithChecklistWarning
+        onClick={handlePublish}
+        isDisabled={!capabilities?.hasPublishAction || isSaving || isUploading}
+        text={text}
+      />
       <TitleMissingDialog
         open={Boolean(showDialog)}
         onIgnore={publish}
