+++ assets/src/edit-story/components/header/buttons/update.js
@@ -27,12 +27,9 @@ import { useFeatures } from 'flagged';
  * Internal dependencies
  */
 import { useStory, useLocalMedia, useHistory, useConfig } from '../../../app';
-import { Outline, Primary } from '../../button';
+import { Outline } from '../../button';
 import { useGlobalKeyDownEffect } from '../../../../design-system';
-import WithTooltip from '../../tooltip';
-import { PRE_PUBLISH_MESSAGE_TYPES } from '../../../app/prepublish';
-import { usePrepublishChecklist } from '../../inspector/prepublish';
-import { ButtonContent, WarningIcon } from './styles';
+import ButtonWithChecklistWarning from './buttonWithChecklistWarning';
 
 function Update() {
   const { isSaving, status, saveStory } = useStory(
@@ -51,11 +48,7 @@ function Update() {
     state: { hasNewChanges },
   } = useHistory();
 
-  const { checklist, refreshChecklist } = usePrepublishChecklist();
-  const {
-    showPrePublishTab,
-    customMetaBoxes: isMetaBoxesFeatureEnabled,
-  } = useFeatures();
+  const { customMetaBoxes: isMetaBoxesFeatureEnabled } = useFeatures();
   const { metaBoxes = {} } = useConfig();
 
   const hasMetaBoxes =
@@ -99,28 +92,13 @@ function Update() {
       );
   }
 
-  const tooltip = showPrePublishTab
-    ? checklist.some(({ type }) => PRE_PUBLISH_MESSAGE_TYPES.ERROR === type) &&
-      __('There are items in the checklist to resolve', 'web-stories')
-    : null;
-
-  const button = (
-    <Primary
-      onPointerEnter={refreshChecklist}
+  return (
+    <ButtonWithChecklistWarning
       onClick={() => saveStory()}
       isDisabled={isSaving || isUploading}
-    >
-      <ButtonContent>
-        {text}
-        {tooltip && <WarningIcon />}
-      </ButtonContent>
-    </Primary>
+      text={text}
+    />
   );
-
-  const wrappedWithTooltip = (
-    <WithTooltip title={tooltip}>{button}</WithTooltip>
-  );
-  return tooltip ? wrappedWithTooltip : button;
 }
 
 export default Update;
